refs #1257 - config.xml and commandline handling. Found (2) bugs where the command line was being overrun by other variables. 1st part of fix, testing is needed.

+++ src/qOrbiter/qOrbiter_src/orbiterwindow.h
@@ -53,7 +53,7 @@ public:
  * \param frameless - set to true of no frame is desired.
  * \param parent
      */
-    explicit orbiterWindow(long deviceid, std::string routerip, bool fullScreen, bool frameless,  QObject *parent = 0);
+    explicit orbiterWindow(int deviceid, std::string routerip, bool fullScreen, bool frameless,  QObject *parent = 0);
     //public members
 
     QString message;
